Tests/LibWeb/Text/expected/hit_testing/wrapped-link-offset-x.txt

rects: 2
offsetX: 5
offsetY: 25
